swapped bit lengths

+++ pyobs_asi/asicamera.py
@@ -211,10 +211,10 @@ class AsiCamera(BaseCamera, ICamera, ICameraWindow, ICameraBinning, IImageFormat
         # decide on image format
         if image_format == asi.ASI_IMG_RAW8:
             shape = [whbi[1], whbi[0]]
-            data = np.frombuffer(buffer, dtype=np.uint16).reshape(shape)
+            data = np.frombuffer(buffer, dtype=np.uint8).reshape(shape)
         elif image_format == asi.ASI_IMG_RAW16:
             shape = [whbi[1], whbi[0]]
-            data = np.frombuffer(buffer, dtype=np.uint8).reshape(shape)
+            data = np.frombuffer(buffer, dtype=np.uint16).reshape(shape)
         elif image_format == asi.ASI_IMG_RGB24:
             shape = [whbi[1], whbi[0], 3]
             data = np.frombuffer(buffer, dtype=np.uint8).reshape(shape)
